Add Resources::Lists#find_by
I want to be consistent around return values; having a create method
that does not return anything when successful is not ideal. After
creating a new list the wrapper will return the created list.

Removed the .to_a call from the Todoable::Client since the wrapper is
the one acting as a proxy between the consumers its better to force the
consumer to be explicit about what data type she/he is going to be
handling.